封装代码/requirements.txt
flask==1.1.0
flask-restful==0.3.8
requests==2.24.0
grpcio==1.32.0
grpcio-tools==1.32.0
protobuf==3.12.2
